+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.50.0 - 2022-03-14
+
+### Added
+
+- A whole new wallets page
+- A new tab in activities page to display network action order history
+
+### Fixed
+
+- Change num token icon from the black one to the blue one
+- Fix activity details cid text overflow. #1167
+- Remove add to contacts checkbox when the receiver already exists in contacts
+
 ## 0.49.1 - 2022-02-25
 
 ### Added